source/devops/requirements.txt
pyshp
flask
flask-cors
python-dotenv
sqlalchemy
psycopg2-binary
pandas
mlflow
pyspark